// src/components/keluarga/form/types.ts
export interface InformasiKeluargaForm {
  rt: string;
  alamat: string;
}

export interface AnggotaForm {
  id: string;

  nama: string;

  jenisKelamin: string;

  tahunKelahiran: string;

  agama: string;

  statusPerkawinan: string;

  pendidikan: string;

  pekerjaan: string;

  statusBantuan: string;

  statusKepemilikanRumah: string;

  statusDalamKeluarga: string;
}